Moved GetNodeColor and GetSegmentColor to Utilities
