+++ Assets/Scripts/Manager.cs
@@ -78,7 +78,7 @@ namespace Rails
                     // draw node
                     var node = Map.Nodes[(y * Size) + x];
                     var pos = GetPosition(node.Id);
-                    Gizmos.color = MapEditorUtils.GetNodeColor(node.Type);
+                    Gizmos.color = Utilities.GetNodeColor(node.Type);
                     Gizmos.DrawSphere(pos, WSSize * 0.1f);
 
                     // draw segments
@@ -86,21 +86,12 @@ namespace Rails
                     for (Cardinal c = 0; c < Cardinal.MAX_CARDINAL; ++c)
                     {
                         var segment = segments[(int)c];
-                        var nextNodeId = new NodeId(-1, -1);
-
-                        switch (c)
-                        {
-                            case Cardinal.S:
-                                {
-                                    nextNodeId = new NodeId(node.Id.X, node.Id.Y + 1);
-                                    break;
-                                }
-                        }
+                        var nextNodeId = Utilities.PointTowards(node.Id, c);
 
                         if (nextNodeId.InBounds)
                         {
                             var nextNode = Map.Nodes[nextNodeId.GetSingleId()];
-                            Gizmos.color = MapEditorUtils.GetSegmentColor(segment.Type);
+                            Gizmos.color = Utilities.GetSegmentColor(segment.Type);
                             Gizmos.DrawLine(pos, GetPosition(nextNode.Id));
                         }
                     }
