--- Assets/Scripts/MapEditor/MapEditorUtils.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fb85e2e5ecabbef4a974154b5f759898
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
