+++ Assets/Scripts/Utilities.cs
@@ -36,6 +36,36 @@ namespace Rails
         /// <param name="towards">The Cardinal being reflected</param>
         /// <returns>The opposite cardinal direction to towards</returns> 
         public static Cardinal ReflectCardinal(Cardinal towards) =>
-            (Cardinal)Mathf.Repeat((int)Cardinal.SW, (int)towards - 3); 
+            (Cardinal)Mathf.Repeat((int)Cardinal.SW, (int)towards - 3);
+
+        /// <summary>
+        /// Returns the respective color associated with the given NodeType.
+        /// </summary>
+        public static Color GetNodeColor(NodeType nodeType)
+        {
+            switch (nodeType)
+            {
+                case NodeType.Clear: return Color.gray;
+                case NodeType.Water: return Color.blue;
+                case NodeType.Mountain: return Color.black;
+                case NodeType.SmallCity: return Color.green;
+                case NodeType.MediumCity: return Color.yellow;
+                case NodeType.MajorCity: return Color.red;
+                default: return Color.white;
+            }
+        }
+
+        /// <summary>
+        /// Returns the respective color associated with the given NodeSegmentType.
+        /// </summary>
+        public static Color GetSegmentColor(NodeSegmentType segmentType)
+        {
+            switch (segmentType)
+            {
+                case NodeSegmentType.None: return Color.white;
+                case NodeSegmentType.River: return Color.blue;
+                default: return Color.white;
+            }
+        }
     }
 }
